Upload Arquyivos

+++ tickets/views.py
@@ -17,9 +17,13 @@ from core.decorators import require_group
 # Create your views here.
 @login_required
 def index_tickets(request):    
-    latest_ticket_list = Ticket.objects.order_by('-created_at')
     template = loader.get_template('tickets/index.html')
+    latest_ticket_list = Ticket.objects.order_by('-created_at')
+    tickets_count_open = Ticket.objects.filter(status=True).count()
+    tickets_count_close = Ticket.objects.filter(status=False).count()
     context = {
+        'tickets_count_open': tickets_count_open,
+        'tickets_count_close': tickets_count_close,
         'latest_ticket_list': latest_ticket_list,
     }
     return HttpResponse(template.render(context, request))
@@ -37,6 +41,7 @@ def create_ticket(request):
             f.status = True #Open
             f.save()
             messages.success(request, 'Adicionado com sucesso!')
+            return HttpResponseRedirect(reverse('tickets:index_tickets'))
     form = TicketForm()
     context['form'] = form
     return render(request, template_name, context)
@@ -51,22 +56,30 @@ def update_ticket(request, ticket_id):
     template_name = 'tickets/create_update.html'
     context = {}
     ticket = get_object_or_404(Ticket, id=ticket_id, user=request.user)
-    if request.method == 'POST':
-        form = TicketForm(request.POST, instance=ticket)
-        if form.is_valid():
-            form.save()
-            messages.success(request, 'Modificado com sucesso!')        
-            return HttpResponseRedirect(reverse('tickets:read_ticket', kwargs={'ticket_id': ticket_id}))
-    form = TicketForm(instance=ticket)
-    context['form'] = form
-    return render(request, template_name, context)
+    if ticket.user == request.user:
+        if request.method == 'POST':
+            form = TicketForm(request.POST, request.FILES, instance=ticket)
+            if form.is_valid():
+                form.save()
+                messages.success(request, 'Modificado com sucesso!')        
+                return HttpResponseRedirect(reverse('tickets:read_ticket', kwargs={'ticket_id': ticket_id}))
+        form = TicketForm(instance=ticket)
+        context['form'] = form
+        return render(request, template_name, context)
+    else:
+        messages.error(request, 'Você não tem permissão para esta ação')        
+        return HttpResponseRedirect(reverse('tickets:read_ticket', kwargs={'ticket_id': ticket_id}))
 
 @login_required
 def delete_ticket(request, ticket_id):
     ticket = get_object_or_404(Ticket, id=ticket_id)
-    ticket.delete()
-    messages.success(request, 'Removido com sucesso!')        
-    return HttpResponseRedirect(reverse('tickets:index_tickets'))
+    if ticket.user == request.user:
+        ticket.delete()
+        messages.success(request, 'Removido com sucesso!')        
+        return HttpResponseRedirect(reverse('tickets:index_tickets'))
+    else:
+        messages.error(request, 'Você não tem permissão para esta ação')        
+        return HttpResponseRedirect(reverse('tickets:read_ticket', kwargs={'ticket_id': ticket_id}))
 
 @login_required
 def close_ticket(request, ticket_id):
@@ -107,13 +120,33 @@ def create_action(request, ticket_id):
 @login_required
 def delete_action(request, action_id, ticket_id):
     action = get_object_or_404(Action, id=action_id)
-    action.delete()
-    messages.success(request, 'Removido com sucesso!')        
-    return HttpResponseRedirect(reverse('tickets:read_ticket', kwargs={'ticket_id': ticket_id}))
+    if action.user == request.user:
+        action.delete()
+        messages.success(request, 'Removido com sucesso!')        
+        return HttpResponseRedirect(reverse('tickets:read_ticket', kwargs={'ticket_id': ticket_id}))
+    else:
+        messages.error(request, 'Você não tem permissão para esta ação')        
+        return HttpResponseRedirect(reverse('tickets:read_ticket', kwargs={'ticket_id': ticket_id}))
 
 @login_required
-def update_action(request, action_id, ticket_id):
-    pass
+def update_action(request, ticket_id, action_id):
+    template_name = 'actions/create_update.html'
+    context = {}
+    action = get_object_or_404(Action, id=action_id, user=request.user)
+    if action.user == request.user:
+        if request.method == 'POST':
+            form = ActionForm(request.POST, request.FILES, instance=action)
+            if form.is_valid():
+                form.save()
+                messages.success(request, 'Modificado com sucesso!')        
+                return HttpResponseRedirect(reverse('tickets:read_ticket', kwargs={'ticket_id': ticket_id}))
+        form = ActionForm(instance=action)
+        context['form'] = form
+        return render(request, template_name, context)
+    else:
+        messages.error(request, 'Você não tem permissão para esta ação')        
+        return HttpResponseRedirect(reverse('tickets:read_ticket', kwargs={'ticket_id': ticket_id}))
+
